Remove warnings from the suite
